Tasks 2 & 3
Completed tasks 2 and 3 (or at least to whatever extent I could within the time limits I set).

Task 2: Made the row generator spawn squares based on the number entered. If checks to make sure it is not null, is an int, and is greater than zero. Tried to spawn them centered in the game screen, but not fully working

Task 3: When mouse held, drew a line from previous position 0.1 sec ago to current position. When mouse let go, printed the total distance since the start of it being held.