Tasks 2 & 3
Forgot to commit task 2 before starting 3. Both are now complete. For task 2 I got the enemy to spawn on a random edge and accelerate in the direction of the player (the player location when they spawned). Task 3 got a random direction and had it move in that direction until it reaches a max distance, and then selects a new direction.