Vector Mechanic Experiment
Began work on the vector mechanic assignment. Goal is to give the ship a baseball bat that can hit shooting stars. Currently have the experiment working so that the balls spawn with a random direction to move in, the bat can reflect the ball's velocity if it is hit on the center point(top left corner of the bat currently). Will continue to iterate and troubleshoot, then try to implement more in depth vector math, including calculating the angle of contact, the speed to add to the ball upon being hit, etc.