Journal 4 Task 1
implemented the first task from the trig challenges. Created a function in the player script, gets a point count and radar radius, then draws line based on it. Green unless enemy is within the radius of the player. I did journal for task 1 before starting task 2, just forgot to commit.